Changes from ATM code review
